+++ src/main/java/me/shepherd23333/projecte/gameObjs/items/TimeWatch.java
@@ -41,6 +41,7 @@ import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 
 import javax.annotation.Nonnull;
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -118,7 +119,7 @@ public class TimeWatch extends ItemPE implements IModeChanger, IBauble, IPedesta
         }
 
         EntityPlayer player = (EntityPlayer) entity;
-        long reqEmc = EMCHelper.removeFractionalEMC(stack, getEmcPerTick(this.getCharge(stack)));
+        BigInteger reqEmc = EMCHelper.removeFractionalEMC(stack, getEmcPerTick(this.getCharge(stack)));
 
         if (!consumeFuel(player, stack, reqEmc, true)) {
             return;
